Increase framerate independence and switch to static images where now required. Add a performance mode
Image animations now run at fullspeed no matter the framerate. The text animators are still framerate dependent, but this is not important as they are needing replacement anyway due to the requirement of non-standard fonts and characters.

Performance mode has been added to the properties set with planned use to facilitate the removal of effects to speed up drawing. Its only current effect is to weaken the logic behind the animation switching to reduce load, although the difference is negligable. Performance metrics show the engine easily reaching 144+ FPS while consuming around 8% of an i5-6200U; while this is favourable to emulation (current melonDS builds use 30% at the same resolution with reduced framerate), there is definitely room for improvement.