Reduce stack control loops, increase hint screen accuracy
The hint screen now follows the action of the base game:
- When opening the hint screen and there are further hints, jump to asking whether the user wants a new hint
- If all hints are unlocked, jump to the last displayed hint

A new method has been added to force the stack to update since it is possible for the stack to be left behind if dynamic context settings are being used.

The puzzle player needs to be optimised, some element(s) is drawing significant amounts of CPU time compared to the room handler.