Add asset unpacker (additional dependencies of ndspy and (optional) pillow)
The ROM can now be used to populate the asset structures for the engine components. LAYTON3 tile unpacking code was backported to allow BGs to be extracted to standard formats (DDS and PNG, latter being much more preferable). The output BG directory size when using PIL is 40MB compared to >150MB.

Current Pillow's DDS importer is quite weak, with the uncompressed DDS side importing exports with incorrect colour masks. The 16-bit variants are not supported, either. To load the correct colours, change the masking code in the DDS plugin to flip the colour arrangement.

The old version of the puzzle player has been removed because the current version is much further ahead.