Add arc/arj animation unpacking capability
Credit to contributors of the Tinke layton plugin for the arj tile reconstruction code.

Animations are now unpacked, with mostly correct alpha and basic timing information. To properly support game timing, the engine must be updated to support per-frame delays. Alpha is measured against the key (0,248,0); the other alpha key (0,240,0) is not taken into account. To actually load the alpha, baking must be disabled in asset_converter as the colorkey is stripped when the alpha is added.