Add RLE decompression and LPCK support for unpacker, start work on making engine start with full game assets
Credit to dsdecmp (FEAT) for flag code. Nestled LPCK containers are not unpacked through this code but this is never used so no big deal

The unpacking script now unpacks most game-critical elements (sound and video will likely not be converted on a platform-independent basis) but needs a refactor because its getting ugly and long for no reason.